Wed, Mar 13, 2024, 1:17 PM +08:00

+++ readability/readability.c
@@ -44,8 +44,8 @@ int main(void)
                 }
     }
 
-    L = (tletters / twords) * 100;
-    S = (tsentences / twords) * 100;
+    L = (tletters / (float) twords) * 100;
+    S = (tsentences / (float) twords) * 100;
 
     float index = 0.0588 * L - 0.296 * S - 15.8;
